Some refactoring for cleaner tests files
